Pull out stm32f7 interrupt support into a separate crate
