+++ interrupture-stm32f7x6/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "interrupture-stm32f7x6"
+version = "0.1.0"
+authors = ["Oliver Scherer <github35764891676564198441@oli-obk.de>"]
+edition = "2018"
+
+[dependencies]
+interrupture = { version = "0.1.0", path = "../interrupture" }
+cortex-m-rt = "0.6.4"
+
+[dependencies.stm32f7]
+version = "0.3.2"
+features = ["stm32f7x6", "rt"]
